Move throttle code to independent source file.  Clean out some dead code within env_posix.cc
